Add and use mockHomebridge in PlatformAccessoryManager test
